stage5/Gimmie_Y0ur_t0ken/requirements.txt
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyJWT==1.7.1
python-http-client==3.2.4
sendgrid==6.1.1
six==1.15.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
requests==2.24.0
gunicorn
